Refactored more of main into other files
