+++ hon_patch_notes_game_bot/user.py
@@ -1,3 +1,4 @@
+#!/usr/bin/python
 class RedditUser:
     def __init__(
         self,
